Dev. ISSUE-0. FINAL DONE

+++ store-service/src/main/java/vvp_company/storeservice/client/dto/ResourceInfoDTO.java
@@ -15,5 +15,5 @@ public class ResourceInfoDTO {
 
     private String description;
 
-    private Integer weightPerUnit;
+    private Double weightPerUnit;
 }
